+++ src/main/java/com/example/controller/UserController.java
@@ -25,7 +25,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public User findById(@PathVariable int id){
+    public User findById(@PathVariable Long id){
         return userService.findById(id);
     }
 }
